+++ examples/smoke.rs
@@ -5,6 +5,8 @@ fn main() {
     println!("{}", humfmt::number(1500000));
     println!("{}", humfmt::number(-12500));
     println!("{}", 1200000000.human_number());
+    println!("{}", humfmt::ordinal(21));
+    println!("{}", 42.human_ordinal());
 
     println!("{}", humfmt::number(999_949));
     println!("{}", humfmt::number(999_950));